move unbind script to instructions
